Updated tests, created tests, and made Teacher_To_TA stuff

+++ TA_Scheduler_App/courseFeatures.py
@@ -15,14 +15,41 @@ class CourseFeatures:
             instructor   = models.CharField(max_length=256, blank=True)
     """
     @staticmethod
-    def create_course(courseName:str,semester:str):
-        course = Course.objects.create(courseName=courseName,semester=semester)
+
+    def create_course(courseName:str):
+        if courseName is None or courseName is '':
+            raise Exception('no')
+
+        a = None
+        try:
+            a = Course.objects.get(courseName=courseName)
+        except Course.MultipleObjectsReturned:
+            raise Exception("something is really wrong")
+        except Course.DoesNotExist:
+            pass
+
+        if a is not None:
+            return False
+
+        course = Course.objects.create(courseName=courseName)
         course.save()
         return course
 
     @staticmethod
-    def create_section(course: Course, sectionCode: str,instructor: str = ""):
-        section = Section.objects.create(course=course, sectionCode=sectionCode,instructor=instructor,)
+
+    def create_section(courseForeignKey:Course, sectionData:str=None):
+        if courseForeignKey is None:
+            return False
+
+        if sectionData is None:
+            return False
+
+        if not isinstance(courseForeignKey,Course):
+            raise Exception('not a course')
+
+        section = Section.objects.create(course=courseForeignKey, sectionCode=sectionData)
+        section.save()
+
         return section
 
     # @staticmethod
@@ -43,6 +70,8 @@ class CourseFeatures:
 
     @staticmethod
     def delete_section(sectionKey:int):
+        if not isinstance(sectionKey,int):
+            return False
         try:
             section = Section.objects.get(pk=sectionKey)
             section.delete()
@@ -61,7 +90,10 @@ class CourseFeatures:
     #         return False
 
     @staticmethod
-    def edit_course(courseKey:int, newCourseName:str="",newSemester:str = ""):
+
+    def edit_course(courseKey:int, newCourseName:str=""):
+        if not isinstance(courseKey,int) or not isinstance(newCourseName,str):
+            return False
         try:
             course = Course.objects.get(pk=courseKey)
         except Course.DoesNotExist:
@@ -76,7 +108,10 @@ class CourseFeatures:
         return course
 
     @staticmethod
-    def edit_section(sectionKey: int, newSectionCode: str = "", newInstructor: str = ""):
+
+    def edit_section(sectionKey:int, newSectionTime:str=None):
+        if not isinstance(sectionKey,int):
+            return False
         try:
             section = Section.objects.get(pk=sectionKey)
         except Section.DoesNotExist:
